🚚 Rename use_case_guide.dart to content_card.dart
